refactor: merch_agents → alchemist_agents にリネーム + Dockerfile 追加
エージェント名（Alchemist）に合わせてパッケージ名を統一。
Dockerfile に COPY 行を追加し、Cloud Run 上の No module named エラーを解消。
