+++ alchemist_agents/__main__.py
@@ -1,0 +1,3 @@
+from alchemist_agents.cli import main
+
+main()
