--- merch_agents/__main__.py
@@ -1,3 +1,0 @@
-from merch_agents.cli import main
-
-main()
